1. Add default policy profile config paramater to cisco_plugins.ini eg. default_policy_profile=profilename for dhcp port. 2. Add physical_network attribute to network profile resource. Change in command line. For type vlan, along with the usual options include --physical_networl PHYS_NET_NAME 3. Add support for vlan/vxlan ranges to be configured via network profile instead of cisco_pluigns.ini

+++ quantum/plugins/cisco/db/n1kv_db_v2.py
@@ -105,29 +105,40 @@ def sync_vlan_allocations(network_vlan_ranges):
             for vlan_range in vlan_ranges:
                 vlan_ids |= set(xrange(vlan_range[0], vlan_range[1] + 1))
 
-            # remove from table unallocated vlans not currently allocatable
+            # add missing allocatable vlans to table
+            for vlan_id in sorted(vlan_ids):
+                try:
+                    alloc = (session.query(n1kv_models_v2.N1kvVlanAllocation).
+                             filter_by(physical_network=physical_network).
+                             filter_by(vlan_id=vlan_id).one())
+                except exc.NoResultFound:
+                    alloc = n1kv_models_v2.N1kvVlanAllocation(physical_network,
+                                                              vlan_id)
+                    session.add(alloc)
+
+def delete_vlan_allocations(network_vlan_ranges):
+    """Delete vlan_allocations for deleted network profile range"""
+
+    session = db.get_session()
+    with session.begin():
+        # process vlan ranges for each physical network separately
+        for physical_network, vlan_ranges in network_vlan_ranges.iteritems():
+            # Determine the set of vlan ids which need to be deleted.
+            vlan_ids = set()
+            for vlan_range in vlan_ranges:
+                vlan_ids |= set(xrange(vlan_range[0], vlan_range[1] + 1))
+
             allocs = (session.query(n1kv_models_v2.N1kvVlanAllocation).
                         filter_by(physical_network=physical_network).
                         all())
             for alloc in allocs:
-                try:
-                    # see if vlan is allocatable
-                    vlan_ids.remove(alloc.vlan_id)
-                except KeyError:
-                    # it's not allocatable, so check if its allocated
+                if alloc.vlan_id in vlan_ids:
                     if not alloc.allocated:
-                        # it's not, so remove it from table
+                        # it's not allocated, so remove it from table
                         LOG.debug("removing vlan %s on physical network "
                                     "%s from pool" %
                                     (alloc.vlan_id, physical_network))
                         session.delete(alloc)
- 
-            # add missing allocatable vlans to table
-            for vlan_id in sorted(vlan_ids):
-                alloc = n1kv_models_v2.N1kvVlanAllocation(physical_network,
-                                                         vlan_id)
-                session.add(alloc)
-
 
 def get_vlan_allocation(physical_network, vlan_id):
     session = db.get_session()
@@ -157,7 +168,7 @@ def reserve_vlan(session, profile):
             physical_network = alloc.physical_network
             alloc.allocated = True
             return (physical_network, segment_type, segment_id, '0.0.0.0')
-    raise q_exc.NoNetworkAvailable()
+        raise q_exc.NoNetworkAvailable()
 
 
 def reserve_vxlan(session, profile):
@@ -259,25 +270,37 @@ def sync_vxlan_allocations(vxlan_id_ranges):
 
     session = db.get_session()
     with session.begin():
-        # remove from table unallocated vxlans not currently allocatable
+        # add missing allocatable vxlans to table
+        for vxlan_id in sorted(vxlan_ids):
+            try:
+                alloc = (session.query(n1kv_models_v2.N1kvVxlanAllocation).
+                         filter_by(vxlan_id=vxlan_id).one())
+            except exc.NoResultFound:
+                alloc = n1kv_models_v2.N1kvVxlanAllocation(vxlan_id)
+                session.add(alloc)
+
+def delete_vxlan_allocations(vxlan_id_ranges):
+    """Delete vxlan_allocations for deleted network profile range"""
+    # determine current configured allocatable vxlans
+    vxlan_ids = set()
+    for vxlan_id_range in vxlan_id_ranges:
+        tun_min, tun_max = vxlan_id_range
+        if tun_max + 1 - tun_min > 1000000:
+            LOG.error("Skipping unreasonable vxlan ID range %s:%s" %
+                      vxlan_id_range)
+        else:
+            vxlan_ids |= set(xrange(tun_min, tun_max + 1))
+
+    session = db.get_session()
+    with session.begin():
         allocs = (session.query(n1kv_models_v2.N1kvVxlanAllocation).all())
         for alloc in allocs:
-            try:
-                # see if vxlan is allocatable
-                vxlan_ids.remove(alloc.vxlan_id)
-            except KeyError:
-                # it's not allocatable, so check if its allocated
+            if alloc.vxlan_id in vxlan_ids:
                 if not alloc.allocated:
                     # it's not, so remove it from table
                     LOG.debug("removing vxlan %s from pool" %
                                 alloc.vxlan_id)
                     session.delete(alloc)
-
-        # add missing allocatable vxlans to table
-        for vxlan_id in sorted(vxlan_ids):
-            alloc = n1kv_models_v2.N1kvVxlanAllocation(vxlan_id)
-            session.add(alloc)
-
 
 def get_vxlan_allocation(vxlan_id):
     session = db.get_session()
@@ -447,7 +470,8 @@ def create_network_profile(profile):
             net_profile = n1kv_models_v2.NetworkProfile(\
                             name=profile['name'],
                             segment_type=profile['segment_type'],
-                            segment_range=profile['segment_range'])
+                            segment_range=profile['segment_range'],
+                            physical_network=profile['physical_network'])
         elif profile['segment_type'] == 'vxlan':
             net_profile = n1kv_models_v2.NetworkProfile(\
                             name=profile['name'],
@@ -473,6 +497,7 @@ def delete_network_profile(id):
         session.delete(profile)
         session.query(n1kv_models_v2.ProfileBinding).filter(\
             n1kv_models_v2.ProfileBinding.profile_id == id).delete()
+    return profile
 
 
 def update_network_profile(id, profile):
@@ -520,14 +545,22 @@ def get_network_profile_by_name(name):
     except exc.NoResultFound:
         return None
 
-def _get_network_profiles():
+def _get_network_profiles(**kwargs):
     """
     Get Network Profiles
 
     :return:
     """
     session = db.get_session()
-    return session.query(n1kv_models_v2.NetworkProfile).all()
+    if "physical_network" in kwargs:
+        try:
+            profiles = session.query(n1kv_models_v2.NetworkProfile).\
+                           filter_by(physical_network=kwargs['physical_network']).all()
+            return profiles
+        except exc.NoResultFound:
+            return None
+    else:
+        return session.query(n1kv_models_v2.NetworkProfile).all()
 
 
 def create_policy_profile(profile):
@@ -613,6 +646,7 @@ def create_profile_binding(tenant_id, profile_id, profile_type):
         binding = n1kv_models_v2.ProfileBinding(profile_type=profile_type,
                         profile_id=profile_id, tenant_id=tenant_id)
         session.add(binding)
+        session.flush()
         return binding
 
 
@@ -690,7 +724,8 @@ class NetworkProfile_db_mixin(object):
                'segment_type': profile['segment_type'],
                'segment_range': profile['segment_range'],
                'multicast_ip_index': profile['multicast_ip_index'],
-               'multicast_ip_range': profile['multicast_ip_range']}
+               'multicast_ip_range': profile['multicast_ip_range'],
+               'physical_network': profile['physical_network']}
         return self._fields(res, fields)
 
     def create_network_profile(self, context, network_profile):
@@ -772,7 +807,7 @@ class NetworkProfile_db_mixin(object):
 
     def _validate_vlan(self, p):
         """Validate if vlan falls within segment boundaries."""
-
+        '''
         seg_min, seg_max = self._get_segment_range(p['segment_range'])
         ranges = conf.CISCO_N1K.network_vlan_ranges
         ranges = ranges.split(',')
@@ -784,6 +819,8 @@ class NetworkProfile_db_mixin(object):
                     msg = _("Vlan out of range")
                     LOG.exception(msg)
                     raise q_exc.InvalidInput(error_message=msg)
+        '''
+        pass
 
     def _validate_vxlan(self, p):
         """
@@ -791,6 +828,7 @@ class NetworkProfile_db_mixin(object):
         :param p:
         :return:
         """
+        '''
         seg_min, seg_max = self._get_segment_range(p['segment_range'])
         ranges = conf.CISCO_N1K.vxlan_id_ranges
         ranges = ranges.split(',')
@@ -812,6 +850,8 @@ class NetworkProfile_db_mixin(object):
             if _validate_ip_address(ip) != None:
                 msg = _("invalid ip address %s" % ip)
                 raise q_exc.InvalidInput(error_message=msg)
+        '''
+        pass
 
     def _validate_segment_range(self, p):
         """
@@ -855,26 +895,27 @@ class NetworkProfile_db_mixin(object):
         :return:
         """
         # self.get_network_profiles(context)
-        profiles = _get_network_profiles()
-        for prfl in profiles:
-            _name = prfl.name
-            _segment_range = prfl.segment_range
-            if p['name'] == _name:
-                msg = _("NetworkProfile name %s already exists" % p['name'])
-                LOG.exception(msg)
-                raise q_exc.InvalidInput(error_message=msg)
-            seg_min, seg_max = self._get_segment_range(p['segment_range'])
-            prfl_seg_min, prfl_seg_max = self._get_segment_range(\
+        profiles = _get_network_profiles(physical_network=p['physical_network'])
+        if profiles:
+            for prfl in profiles:
+                _name = prfl.name
+                _segment_range = prfl.segment_range
+                if p['name'] == _name:
+                    msg = _("NetworkProfile name %s already exists" % p['name'])
+                    LOG.exception(msg)
+                    raise q_exc.InvalidInput(error_message=msg)
+                seg_min, seg_max = self._get_segment_range(p['segment_range'])
+                prfl_seg_min, prfl_seg_max = self._get_segment_range(\
                                             _segment_range)
-            if (((seg_min >= prfl_seg_min) and
-                 (seg_min <= prfl_seg_max)) or
-                ((seg_max >= prfl_seg_min) and
-                 (seg_max <= prfl_seg_max)) or
-                ((seg_min <= prfl_seg_min) and
-                 (seg_max >= prfl_seg_max))):
-                msg = _("segment range overlaps with another profile")
-                LOG.exception(msg)
-                raise q_exc.InvalidInput(error_message=msg)
+                if (((seg_min >= prfl_seg_min) and
+                     (seg_min <= prfl_seg_max)) or
+                    ((seg_max >= prfl_seg_min) and
+                     (seg_max <= prfl_seg_max)) or
+                    ((seg_min <= prfl_seg_min) and
+                     (seg_max >= prfl_seg_max))):
+                    msg = _("segment range overlaps with another profile")
+                    LOG.exception(msg)
+                    raise q_exc.InvalidInput(error_message=msg)
 
 
 class PolicyProfile_db_mixin(object):
@@ -1045,5 +1086,4 @@ class PolicyProfile_db_mixin(object):
         tenant_id = tenant_id or n1kv_models_v2.TENANT_ID_NOT_SET
         if not self._policy_profile_exists(profile_id):
             create_policy_profile(profile)
-        if not _profile_binding_exists(tenant_id, profile['id'], 'policy'):
-            create_profile_binding(tenant_id, profile['id'], 'policy')
+        create_profile_binding(tenant_id, profile['id'], 'policy')
